Move ObjectKind to its own submodule.
